[#6] refactor: 알림 관련(레포지토리) 추상화
