+++ module-core/src/main/java/com/soundie/notification/service/NotificationService.java
@@ -2,7 +2,7 @@ package com.soundie.notification.service;
 
 import com.soundie.notification.domain.Notification;
 import com.soundie.notification.dto.GetNotificationResDto;
-import com.soundie.notification.repository.NotificationRepository;
+import com.soundie.notification.repository.MemoryNotificationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +12,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class NotificationService {
 
-    private final NotificationRepository notificationRepository;
+    private final MemoryNotificationRepository notificationRepository;
 
     public GetNotificationResDto readNotificationList(Long memberId) {
         List<Notification> findNotifications = notificationRepository.findNotificationsByMemberId(memberId);
